+++ lib/src/framework/hover_framework.dart
@@ -2,4 +2,3 @@ library hover_framework;
 
 export 'hover_app.dart';
 export 'core/hover_page.dart';
-export 'core/hover_swapper.dart';