+++ lib/src/widgets/forms/hover_email_sign_in_form.dart
@@ -1,12 +1,12 @@
 import 'package:flutter/widgets.dart';
-import 'fields/email_field.dart';
-import 'fields/password_field.dart';
+import 'fields/hover_email_field.dart';
+import 'fields/hover_password_field.dart';
 
 import 'base/custom_form.dart';
 
-class EmailSignInForm extends CustomForm {
-  EmailSignInForm({
-    @required Future<String> Function(String email, String password) onSubmit,
+class HoverEmailSignInForm extends CustomForm {
+  HoverEmailSignInForm({
+    @required Function(String email, String password) onSubmit,
     @required String formName,
     Color submitTextColor,
     Color submitButtonColor,
@@ -14,13 +14,13 @@ class EmailSignInForm extends CustomForm {
           formName: formName,
           title: "Sign In",
           fields: [
-            EmailField(initialValue: "ernpao@gmail.com"),
-            PasswordField(initialValue: "password"),
+            HoverEmailField(),
+            PasswordField(),
           ],
           onSubmit: (Map<String, String> map) {
-            String email = map[EmailField.fieldName];
+            String email = map[HoverEmailField.fieldName];
             String password = map[PasswordField.fieldName];
-            return onSubmit(email, password);
+            onSubmit(email, password);
           },
           submitTextColor: submitTextColor,
           submitColor: submitButtonColor,