+++ lib/src/widgets/hover_widgets.dart
@@ -1,18 +1,26 @@
 library hover_widgets;
 
 // Basic buttons
-export 'buttons/call_to_action_button.dart';
-export 'buttons/circle_icon_button.dart';
+export 'buttons/hover_call_to_action_button.dart';
+export 'buttons/hover_circle_icon_button.dart';
 
 //Social Media buttons
-export 'buttons/social/facebook_sign_in_button.dart';
-export 'buttons/social/google_sign_in_button.dart';
+export 'buttons/social/hover_facebook_sign_in_button.dart';
+export 'buttons/social/hover_google_sign_in_button.dart';
 
 //Custom Widgets
-export 'custom/full_page_spinner.dart';
-export 'custom/rectangular_image.dart';
-export 'custom/square_image.dart';
-export 'custom/image_slider.dart';
+export 'custom/hover_full_page_spinner.dart';
+export 'custom/hover_rectangular_image.dart';
+export 'custom/hover_square_image.dart';
+export 'custom/hover_image_slider.dart';
+export 'custom/hover_search_bar.dart';
 
 //Typography
 export 'typography/typography.dart';
+
+//Form Widgets
+export 'forms/hover_email_sign_in_form.dart';
+export 'forms/hover_email_sign_up_form.dart';
+export 'forms/hover_password_reset_form.dart';
+export 'forms/fields/hover_email_field.dart';
+export 'forms/fields/hover_password_field.dart';